+++ cmd/go.mod
@@ -1,0 +1,18 @@
+module github.com/google/osv-scanner/cmd
+
+go 1.21.5
+
+require (
+	github.com/urfave/cli/v2 v2.26.0
+)
+
+require (
+	github.com/BurntSushi/toml v1.3.2 // indirect
+	github.com/google/osv-scanner v1.5.0 // indirect
+	github.com/package-url/packageurl-go v0.1.2 // indirect
+	golang.org/x/mod v0.14.0 // indirect
+	golang.org/x/sync v0.5.0 // indirect
+	golang.org/x/sys v0.15.0 // indirect
+	golang.org/x/term v0.15.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
